Clone Babun source in target from correct origin repository
For the copy of the Babun source in the target repository, instead of
always cloning from the Babun GitHub repository, clone it from whichever
repository is set as the origin of the local repository used for
building. This ensures that any commits not in upstream are included in
the build.