Respect remote repository setting when checking for update
Use the remote repository specified by the local Babun source when
checking for updates. This makes sure that any custom commits are kept
when updating.

'cygwin.version' is added to the root directory to check for updated
Cygwin. This gives the benefit of limited control over Cygwin updates
that's branch-specific; e.g. new Cygwin updates can be tested on the
branch first before changing 'cygwin.version' to allow all installations
to update.